refactor: architecture cleanups batch 2 — DeviceTokenService, OAuthRepository move, VoteService/PollService to CoreComponents
- Extract DeviceTokenService from DeviceRegistrationApi (service layer for
  device token registration/deregistration with validation and audit logging)
- Move OAuthRepository, OAuthConnection, OAuthUserInfo from platform-security
  to platform-core (fixes inverted persistence→security dependency)
- Move VoteService and PollService creation from WebFactory to CoreComponents
  (consistent with MessageService/ContactService pattern)
