+++ platform-persistence-jdbi/src/test/kotlin/io/github/rygel/outerstellar/platform/persistence/JdbiOAuthRepositoryTest.kt
@@ -1,6 +1,5 @@
 package io.github.rygel.outerstellar.platform.persistence
 
-import io.github.rygel.outerstellar.platform.security.OAuthConnection
 import java.util.UUID
 import kotlin.test.Test
 import kotlin.test.assertEquals
