+++ platform-web/src/test/kotlin/io/github/rygel/outerstellar/platform/web/Stubs.kt
@@ -4,12 +4,12 @@ import io.github.rygel.outerstellar.platform.model.DeviceToken
 import io.github.rygel.outerstellar.platform.model.Session
 import io.github.rygel.outerstellar.platform.persistence.DeviceTokenRepository
 import io.github.rygel.outerstellar.platform.persistence.MessageCache
+import io.github.rygel.outerstellar.platform.persistence.OAuthConnection
+import io.github.rygel.outerstellar.platform.persistence.OAuthRepository
 import io.github.rygel.outerstellar.platform.persistence.OutboxEntry
 import io.github.rygel.outerstellar.platform.persistence.OutboxRepository
 import io.github.rygel.outerstellar.platform.persistence.OutboxStatus
 import io.github.rygel.outerstellar.platform.persistence.SessionRepository
-import io.github.rygel.outerstellar.platform.security.OAuthConnection
-import io.github.rygel.outerstellar.platform.security.OAuthRepository
 import java.time.Instant
 import java.util.UUID
